+++ cmms/src/main/java/br/com/cmms/cmms/model/Peca.java
@@ -42,7 +42,7 @@ public class Peca {
 
     // 🔹 GETTERS / SETTERS
     public Long getId() {
-        return id;
+        return (Long) id;
     }
 
     public String getNome() {
